chat with cv/requirements.txt
google-genai>=1.0.0
chromadb>=0.4.0
pdfplumber>=0.10.0
python-docx>=1.0.0
streamlit>=1.30.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
